[BUGFIX] Send the correct information to Beebox for file download.
The files can be stored inside folders on Beebox (e.g. de/filename.xml)
and every file can have a locale (de_CH). At the moment the Downloader
assumes the files is stored inside a folder and prefixes the filename
with the locale (DE) but does not provide the locales for the API.

This commit changes it. To not break existing functionality it
introduces a new settings `remoteFilename` and passes this and the
locale to Beebox.

ATTENTION: This must committed together with the corrospending change in
ext:localizer_beebox