[TASK] Use a "root" editorconfig
localizer is a project, and therefore should have its own root editorconfig file.
Otherwise when you work on localizer the config will differ depending on the
parent project it is loaded in.

I have copied .editorconfig from the current TYPO3 master which will become v11.
It should be kept in sync with the latest supported TYPO3 version.