[TASK] Throw exception when export job fails
The result of the export job, which is done by ext:l10nmgr is not taken
into account in the scheduler task. Therefore the GUI shows the user a
"green" success messsage even if an error occured. To be able to see
that and which error occured, we added an exception.

Disadvantages:

In develoment mode the exception is shown instead of a nice flash
message. In production mode, the exception is hidden and an fatal error
occures but the exception is logged in TYPO3 log.