[BUGFIX] Fix mapping from Beebox locale to TYPO3 internal ISO2
We switch from asking the `static_language` table for ISO2 code to the new
Site object, which contains all configured languages.

1. The method failed mapping from Beebox to TYPO3 by
languages like spanish when the Beebox locale was only two letters (es).
2. Beebox uses dashes for locales (de-CH) and TYPO3
uses underscores (de_CH) and the mapping was done wrong.
3. The table `static_languages` does not contain values for
the field `lg_collate_locale`, which will result in a lot of "Missing
locales errors" in BE, which can't be fixed.